+++ graphql_client_codegen/src/lib.rs
@@ -23,9 +23,9 @@ extern crate syn;
 extern crate quote;
 
 use proc_macro2::TokenStream;
-use syn::Visibility;
 
 mod codegen;
+mod codegen_options;
 /// Deprecation-related code
 pub mod deprecation;
 mod introspection_response;
@@ -47,10 +47,10 @@ mod shared;
 mod unions;
 mod variables;
 
-use heck::SnakeCase;
-
 #[cfg(test)]
 mod tests;
+
+pub use codegen_options::GraphQLClientCodegenOptions;
 use proc_macro2::{Ident, Span};
 
 type CacheMap<T> =
@@ -62,36 +62,16 @@ lazy_static! {
         CacheMap::default();
 }
 
-/// Used to configure code generation.
-#[derive(Clone)]
-pub struct GraphQLClientDeriveOptions {
-    /// Name of the operation we want to generate code for. If it does not match, we use all queries.
-    pub operation_name: Option<String>,
-    /// The name of implemention target struct.
-    pub struct_name: Option<String>,
-    /// The module that contains queries.
-    pub module_name: Option<String>,
-    /// Comma-separated list of additional traits we want to derive.
-    pub additional_derives: Option<String>,
-    /// The deprecation strategy to adopt.
-    pub deprecation_strategy: Option<deprecation::DeprecationStrategy>,
-    /// target module visibility.
-    pub module_visibility: Visibility,
-}
-
 /// Generates the code for a Rust module given a query, a schema and options.
 pub fn generate_module_token_stream(
     query_path: std::path::PathBuf,
     schema_path: &std::path::Path,
-    options: Option<GraphQLClientDeriveOptions>,
+    options: GraphQLClientCodegenOptions,
 ) -> Result<TokenStream, failure::Error> {
-    let options = options.unwrap();
-
-    let module_visibility = options.module_visibility.clone();
-    let response_derives = options.additional_derives.clone();
+    let response_derives = options.additional_derives();
 
     // The user can determine what to do about deprecations.
-    let deprecation_strategy = options.deprecation_strategy.clone().unwrap_or_default();
+    let deprecation_strategy = options.deprecation_strategy();
 
     // We need to qualify the query with the path to the crate it is part of
     let (query_string, query) = {
@@ -107,16 +87,12 @@ pub fn generate_module_token_stream(
     };
 
     // Determine which operation we are generating code for. This will be used in operationName.
-    let operations = if options.operation_name.is_some() {
-        let op = codegen::select_operation(&query, &(options.operation_name.clone().unwrap()));
-        if op.is_some() {
-            vec![op.unwrap()]
-        } else {
-            codegen::all_operations(&query)
-        }
-    } else {
-        codegen::all_operations(&query)
-    };
+    let operations = options
+        .operation_name
+        .as_ref()
+        .and_then(|operation_name| codegen::select_operation(&query, &operation_name))
+        .map(|op| vec![op])
+        .unwrap_or_else(|| codegen::all_operations(&query));
 
     let schema_extension = schema_path
         .extension()
@@ -148,24 +124,14 @@ pub fn generate_module_token_stream(
                     };
     let schema = schema::Schema::from(&parsed_schema);
 
-    let struct_name = if options.struct_name.is_some() {
-        Some(Ident::new(
-            options.struct_name.clone().unwrap().as_str(),
-            Span::call_site(),
-        ))
-    } else {
-        None
-    };
+    let struct_name: Option<Ident> = options
+        .struct_name
+        .as_ref()
+        .map(|struct_name| Ident::new(struct_name, Span::call_site()));
 
-    let module_name = Ident::new(
-        options
-            .module_name
-            .as_ref()
-            .unwrap_or_else(|| options.operation_name.as_ref().unwrap())
-            .to_snake_case()
-            .as_str(),
-        Span::call_site(),
-    );
+    let module_name = options
+        .module_name_ident()
+        .ok_or_else(|| format_err!("Could not infer a name for the generated module."))?;
 
     let operation_count = operations.len();
 
@@ -175,10 +141,10 @@ pub fn generate_module_token_stream(
 
     for operation in &operations {
         let schema_output = codegen::response_for_query(
-            &schema.clone(),
-            &query.clone(),
+            &schema,
+            &query,
             &operation,
-            response_derives.clone(),
+            response_derives,
             deprecation_strategy.clone(),
             multiple_operations,
         )?;
@@ -187,9 +153,9 @@ pub fn generate_module_token_stream(
     }
 
     let result = build_module_token_stream(
-        &module_visibility,
+        &options,
         &module_name,
-        &struct_name,
+        struct_name.as_ref(),
         &query_string,
         schema_and_operations,
     );
@@ -198,19 +164,19 @@ pub fn generate_module_token_stream(
 }
 
 fn build_module_token_stream(
-    module_visibility: &syn::Visibility,
+    options: &GraphQLClientCodegenOptions,
     module_name: &Ident,
-    struct_name: &Option<Ident>,
+    struct_name: Option<&Ident>,
     query_string: &str,
     schema_and_operations: Vec<(TokenStream, Ident, &str)>,
 ) -> TokenStream {
-    let mut schema_token_streams = vec![];
-    let mut trait_token_streams = vec![];
+    let mut schema_token_streams = Vec::with_capacity(schema_and_operations.len());
+    let mut trait_token_streams = Vec::with_capacity(schema_and_operations.len());
     let multiple_operations = schema_and_operations.len() > 1;
     for (schema_output, operation_name, operation_name_literal) in schema_and_operations {
         let (schema_token_stream, trait_token_stream) = build_query_struct_token_stream(
             &module_name,
-            struct_name.clone(),
+            struct_name,
             &schema_output,
             &operation_name,
             operation_name_literal,
@@ -221,21 +187,32 @@ fn build_module_token_stream(
     }
 
     merge_with_common_token_stream(
-        &module_visibility,
+        &options,
         &module_name,
         query_string,
-        schema_token_streams,
-        trait_token_streams,
+        &schema_token_streams,
+        &trait_token_streams,
     )
 }
 
 fn merge_with_common_token_stream(
-    module_visibility: &syn::Visibility,
+    options: &GraphQLClientCodegenOptions,
     module_name: &Ident,
     query_string: &str,
-    schema_token_streams: Vec<TokenStream>,
-    trait_token_streams: Vec<TokenStream>,
+    schema_token_streams: &[TokenStream],
+    trait_token_streams: &[TokenStream],
 ) -> TokenStream {
+    let module_visibility = options.module_visibility();
+
+    // Force cargo to refresh the generated code when the query file changes.
+    let query_include = options
+        .query_file()
+        .map(|path| {
+            let path = path.to_str();
+            quote!(const __QUERY_WORKAROUND: &str = include_str!(#path))
+        })
+        .unwrap_or_else(|| quote! {});
+
     quote!(
         #module_visibility mod #module_name {
             #![allow(non_camel_case_types)]
@@ -245,6 +222,9 @@ fn merge_with_common_token_stream(
             use serde;
 
             pub const QUERY: &'static str = #query_string;
+
+            #query_include;
+
             #(#schema_token_streams)*
         }
         #(#trait_token_streams)*
@@ -253,19 +233,15 @@ fn merge_with_common_token_stream(
 
 fn build_query_struct_token_stream(
     module_name: &Ident,
-    struct_name: Option<Ident>,
+    struct_name: Option<&Ident>,
     schema_output: &TokenStream,
     operation_name: &Ident,
     operation_name_literal: &str,
     multiple_operations: bool,
 ) -> (TokenStream, TokenStream) {
-    let struct_name = if struct_name.is_some() {
-        struct_name.unwrap()
-    } else {
-        operation_name.clone()
-    };
+    let struct_name = struct_name.unwrap_or(operation_name);
 
-    let (respons_data_struct_name, variables_struct_name) = if multiple_operations {
+    let (response_data_struct_name, variables_struct_name) = if multiple_operations {
         (
             Ident::new(
                 format!("{}ResponseData", operation_name_literal).as_str(),
@@ -290,7 +266,7 @@ fn build_query_struct_token_stream(
     let trait_token = quote!(
         impl ::graphql_client::GraphQLQuery for #struct_name {
             type Variables = #module_name::#variables_struct_name;
-            type ResponseData = #module_name::#respons_data_struct_name;
+            type ResponseData = #module_name::#response_data_struct_name;
 
             fn build_query(variables: Self::Variables) -> ::graphql_client::QueryBody<Self::Variables> {
                 ::graphql_client::QueryBody {
@@ -306,10 +282,11 @@ fn build_query_struct_token_stream(
 }
 
 fn read_file(path: &::std::path::Path) -> Result<String, failure::Error> {
+    use std::fs;
     use std::io::prelude::*;
 
     let mut out = String::new();
-    let mut file = ::std::fs::File::open(path).map_err(|io_err| {
+    let mut file = fs::File::open(path).map_err(|io_err| {
         let err: failure::Error = io_err.into();
         err.context(format!(
             r#"